+++ packages/core/src/types/plugins/PlatePluginKey.ts
@@ -5,7 +5,7 @@ export type PluginKey = string;
 
 export interface PlatePluginKey {
   /**
-   * Unique key to store the plugins by key.
+   * Property used by Plate to store the plugins by key in `editor.pluginsByKey`.
    */
   key?: PluginKey;
 }